+++ lib/webserver/routes.py
@@ -13,39 +13,31 @@ def setup_routes(app, cors):
         "/api/series/{series_name}/run", ApiHandlers.run_job_for_series)
     app.router.add_get(
         "/api/series/{series_name}/state/{pool_id}/{job_type_id}",
-        ApiHandlers.query_series_state)
-    app.router.add_get(
-        "/api/config/series", ApiHandlers.get_series_configs)
-    app.router.add_post(
-        "/api/config/series", ApiHandlers.add_series_config)
-    app.router.add_delete(
-        "/api/config/series/{rid}", ApiHandlers.remove_series_config)
-    app.router.add_put(
-        "/api/config/series/{rid}/static",
-        ApiHandlers.update_series_config_static)
-    app.router.add_put(
-        "/api/config/series/{rid}", ApiHandlers.update_series_config)
+        ApiHandlers.query_series_state,
+        allow_head=False)
     app.router.add_get(
         "/api/enodo/module", ApiHandlers.get_possible_analyser_modules,
         allow_head=False)
     app.router.add_get(
-        "/api/enodo/output/{output_type}", ApiHandlers.get_enodo_outputs)
+        "/api/enodo/output/{output_type}", ApiHandlers.get_enodo_outputs,
+        allow_head=False)
     app.router.add_delete(
         "/api/enodo/output/{output_type}/{output_id}",
         ApiHandlers.remove_enodo_output)
     app.router.add_post(
         "/api/enodo/output/{output_type}", ApiHandlers.add_enodo_output)
     app.router.add_get(
-        "/api/enodo/stats", ApiHandlers.get_enodo_stats)
+        "/api/enodo/stats", ApiHandlers.get_enodo_stats, allow_head=False)
 
     app.router.add_get(
-        "/api/worker/{pool_id}", ApiHandlers.get_workers)
+        "/api/worker/{pool_id}", ApiHandlers.get_workers, allow_head=False)
     app.router.add_post(
         "/api/worker/{pool_id}", ApiHandlers.add_worker)
     app.router.add_delete(
         "/api/worker/{pool_id}/{job_type_id}", ApiHandlers.delete_worker)
     app.router.add_get(
-        "/api/worker/stats/{pool_id}", ApiHandlers.query_worker_stats)
+        "/api/worker/stats/{pool_id}", ApiHandlers.query_worker_stats,
+        allow_head=False)
 
     # Add internal api routes
     app.router.add_get(
@@ -55,8 +47,6 @@ def setup_routes(app, cors):
     app.router.add_get(
         "/api/enodo/status", ApiHandlers.get_siridb_enodo_status,
         allow_head=False)
-    app.router.add_get(
-        "/api/enodo/log", ApiHandlers.get_event_log, allow_head=False)
     app.router.add_get(
         "/api/enodo/clients", ApiHandlers.get_connected_clients,
         allow_head=False)